+++ gui/src/main/index.ts
@@ -180,6 +180,12 @@ class ApplicationMain {
   );
 
   public run() {
+    // Remove window animations to combat window flickering when opening window. Can be removed when
+    // this issue has been resolved: https://github.com/electron/electron/issues/12130
+    if (process.platform === 'win32') {
+      app.commandLine.appendSwitch('wm-window-animations-disabled');
+    }
+
     // Since electron's GPU blacklists are broken, GPU acceleration won't work on older distros
     if (process.platform === 'linux') {
       app.commandLine.appendSwitch('--disable-gpu');